+++ samples/fn.js
